sgtm_cloud_run_updater/requirements.txt
google-cloud-run==0.3.1  
google-cloud-functions-framework==1.0.2  
requests==2.28.1 
